requirements.txt
absl-py==2.2.2
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albumentations==0.5.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
anytree==2.13.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
boto3==1.37.34
botocore==1.37.34
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmeel==0.57.3
cmeel-assimp==5.4.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.3
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
cmeel-zlib==1.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dash==3.0.3
dataclasses-json==0.6.7
decorator==4.4.2
Deprecated==1.2.18
dex_retargeting==0.4.6
diffusers==0.30.0
dill==0.3.9
docker-pycreds==0.4.0
easydict==1.9
egl_probe==1.0.2
eigenpy==3.5.1
einops==0.8.1
embreex==2.17.7.post6
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.14.0
Flask==3.0.3
flatbuffers==25.2.10
flatdict==4.0.1
fonttools==4.57.0
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.0.0
graphviz==0.20.3
grpcio==1.71.0
gym==0.23.1
gym-notices==0.0.8
gymnasium==1.1.1
h5py==3.13.0
hidapi==0.14.0.post2
hpp-fcl==2.4.4
huggingface-hub==0.29.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib_metadata==8.6.1
iopath==0.1.10
ipycanvas==0.13.3
ipyevents==2.0.2
ipython==8.35.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jaxtyping==0.2.19
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.14.3
llvmlite==0.44.0
loop-rate-limiters==1.1.0
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nlopt==2.7.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
omniverse-kit==106.1.0.140981
onnx==1.16.1
onnxruntime-gpu==1.20.1
open3d==0.19.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
osqp==0.6.7.post3
oss2==2.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pillow==11.0.0
pin==2.7.0
pin-pink==3.1.0
platformdirs==4.3.7
plotly==6.0.1
plyfile==1.1
pooch==1.8.2
portalocker==3.1.1
prettytable==3.3.0
proglog==0.1.11
promise==2.3
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pygame==2.1.0
pyglet==1.5.31
pygltflib==1.16.3
Pygments==2.19.1
pyhocon==0.3.57
PyMatting==1.1.13
pymeshlab==2023.12.post2
PyOpenGL==3.1.0
pyparsing==3.2.3
pyperclip==1.9.0
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-fcl==0.7.0.8
pytorch-lightning==1.9.0
pytorch-msssim==1.0.0
pytransform3d==3.14.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==24.0.1
qdldl==0.1.7.post5
qpsolvers==4.6.0
quadprog==0.1.13
referencing==0.36.2
regex==2024.11.6
rembg==2.0.65
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
segment-anything==1.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shapely==2.1.0
shortuuid==1.0.13
six==1.17.0
skrl==1.4.3
smmap==5.0.2
sniffio==1.3.1
stable_baselines3==2.6.0
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tabulate==0.9.0
taichi==1.6.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==0.4.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.37.2
trimesh==4.6.6
triton==3.1.0
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
usd-core==23.5
wandb==0.12.21
warp-lang==1.7.0
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
xatlas==0.0.10
yapf==0.43.0
yarl==1.19.0
zipp==3.21.0
msgpack==1.1.1
slangtorch==1.3.4